requirements.txt
coinbasepro
python-dotenv